+++ samplesheets/tests/test_rendering.py
@@ -66,7 +66,7 @@ class TestTableBuilder(TestRenderingBase):
         assert_row_length(tables['study'])
 
         # Sources
-        table_sources = get_column_set(tables['study'], 0)
+        table_sources = get_column_set(tables['study'], 1)
         db_sources = set(GenericMaterial.objects.filter(
             study=self.study, item_type='SOURCE').values_list(
             'name', flat=True))